pyproject.toml
[project]
name = "pqc"
description = "Process Quality Control for CMS outer tracker"
readme = "README.md"
authors = [
    {name = "Bernhard Arnold", email = "bernhard.arnold@oeaw.ac.at"},
]
classifiers = [
    "Framework :: PyQt5",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.9"
dependencies = [
    "comet @ git+https://github.com/hephy-dd/comet.git@0.13.2",
    "analysis-pqc @ git+https://github.com/hephy-dd/analysis-pqc.git@0.8.2",
    "PyVISA==1.14.1",
    "PyVISA-py==0.7.2",
    "PyQt5==5.15.11",
    "PyQtChart==5.15.7",
    "PyYAML==6.0.2",
    "jsonschema==3.2.0",
    "flask==3.1.0",
    "waitress==3.0.2",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/hephy-dd/comet-pqc/"
Documentation = "https://hephy-dd.github.io/comet-pqc/"

[project.scripts]
pqc = "pqc.__main__:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "pqc.__version__"}

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"pqc.assets.config.chuck" = ["*.yaml"]
"pqc.assets.config.sample" = ["*.yaml"]
"pqc.assets.config.sequence" = ["*.yaml"]
"pqc.assets.icons" = ["*.svg", "*.png", "*.ico"]
"pqc.assets.schema" = [
    "chuck.yaml",
    "sample.yaml",
    "sequence.yaml",
]
